cleaned up in naming and code
