+++ src/Dan.Plugin.Kartverket/Dan.Plugin.Kartverket.csproj
@@ -11,6 +11,7 @@
     <PackageReference Include="Azure.Identity" Version="1.17.1" />
     <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.8.0" />
     <PackageReference Include="Dan.Common" Version="1.11.0" />
+    <PackageReference Include="Dapper" Version="2.1.72" />
     <PackageReference Include="MaxRev.Gdal.Core" Version="3.12.2.472" />
     <PackageReference Include="Microsoft.Azure.Core.NewtonsoftJson" Version="2.0.0" />
     <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.3.0" />
